+++ packages/tsl-dx/src/index.ts
@@ -1,4 +1,5 @@
 export { noDuplicateExports } from "./rules/no-duplicate-exports";
 export { noDuplicateImports } from "./rules/no-duplicate-imports";
 export { noMultilineTemplateExpressionWithoutAutoDedent } from "./rules/no-multiline-template-expression-without-auto-dedent";
+export { noUnsafeAs } from "./rules/no-unsafe-as";
 export { nullish } from "./rules/nullish";